Tweak deploy workflow to support different envs
The existing Github workflow only had to deploy to dev namespace, but we want to deploy to production next.

This PR makes a few tweaks to support this. For now, the main change is reorganising the manifest files into separate directories, so we have one for `development` and soon we will have another one for `production`, along with the changes to the workflow steps.