Setup error pages and health check endpoint
Configure nginx to serve 404 and 50x error pages. The copy on these pages can be changed if needed.

Expose a health endpoint to be used (in a separate PR) by kubernetes probes to ensure the pods are healthy and handle zero-downtime deploys.

Finally, use the `Dockerfile` container also in `docker-compose` (same nginx image and config in both places for consistency).